extract common utils to a separate file
