+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './browser-parser-utils.js';
 export * from './transformer-utils.js';
+export * from './common-utils.js';
